refactor(root-cms): rename ai-chat.ts to ai.ts
The module now holds the full set of AI features (chat, edit-with-AI,
diff summaries, publish messages, translations, alt text, image
generation), not just chat — the `-chat` suffix is misleading.

Renames the file plus its test, updates imports in api.ts, app.tsx and
plugin.ts, and refreshes the file header doc comment.
